syncing...

+++ Platform.h
@@ -17,12 +17,12 @@
 #endif
 
 #if __cplusplus < 201103L && !defined(_WIN32)
-typedef signed char int8_t;
+typedef char int8_t;
 typedef short int16_t;
-typedef int int32_t;
+typedef long int32_t;
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
-typedef unsigned int uint32_t;
+typedef unsigned long uint32_t;
 #ifdef _AMIGA
 typedef unsigned char bool;
 #endif
